+++ tools/dat-dumper/src/main.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 mod codec;
 mod dat_record;
 mod validate;